scpi: Add alternative legacy structures, functions and macros
In order to support the legacy SCPI protocol variant, add back the structures
and macros that varies against the final specification.
Add indirection table for legacy commands.
Add bitmap field for channel selection
Add support for legacy in scpi_send_message, handle_remote_msg and tx_prepare.
There is a drawback in the mailbox reply path since the Amlogic SCP
firmware reply is always 1 and this value cannot be used to recover the
command in the rx pending list, so we assume the command is the head of the
list and we ensure sequential command sending with a separate mutex.