clk: meson-gxbb: Add GXL/GXM GP0 Variant
The clock tree in the Amlogic GXBB and GXL/GXM SoCs is shared, bus the GXL/GXM
SoCs embeds a different GP0 PLL, amd so needs different parameters and we
only have a reduced rate table.

This patch add the GXL GP0 variant, and adds a GXL DT compatible in order
to use the GXL GP0 PLL instead of the GXBB specific one.