WIP: meson: vdec: Map the Firmware buf idx <-> VB2 buf idx
The current code assumes the vb2 indexes are in order (0 -> n) when
iterating over them. This is very wrong and, in the case where they are
not in the perfect order, will generate mismatches between the firmware
internal buffer indexes and the VB2 ones.

This fixes this bad assumption by explicitely mapping the buffer index.