dt-bindings: phy: meson-g12a-usb3-pcie-phy: convert to yaml
Now that we have the DT validation in place, let's convert the device tree
bindings for the Amlogic G12A USB3 + PCIE Combo PHY over to a YAML schemas.

While the original phy bindings specifies phy-supply as required,
the examples and implementations makes it optional, thus phy-supply
is not present in the properties and required lists.

Reviewed-by: Rob Herring <robh@kernel.org>