FROMGIT: arm64: dts: meson-g12a-x96-max: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit 9a9ffc69901949e3fdd1a8749e544094cb18b899)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>