FROMGIT: arm64: dts: meson-g12b-khadas-vim3: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit ec9037c040411413b84d18224e4de5395fbd2e49)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>