FROMGIT: arm64: dts: meson-gx-p23x-q20x: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit 86b8eaa23ddc62eb58e5a0c9d0a2243b12f52261)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>