FROMGIT: arm64: dts: meson-gxbb-nexbox-a95x: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit 42d7815c2200e4808f8e2e1c4f9c37bf19776fbc)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>