FROMGIT: arm64: dts: meson-gxbb-wetek: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit 6b697024f7bc3122a3185d9e2b388dac4fb5a7c7)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>