arm64: dts: amlogic: meson-sm1-bananapi-m5: set bias on ethernet PHY irq pin
The irq line doesn't have a pull-up/down nor the GPIOZ_14 pin according
to the datasheet of the S905X4 SoC.

Without this, the IRQ line stays at low level triggering an IRQ storm.

This adds a pull-up and disables output on the GPIOZ_14 pin to have
correct behavior for the line.

Reported-by: Christian Hewitt <christianshewitt@gmail.com>
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>