drm/msm/disp: dpu1: Add MSM8996 support
Add support for MSM8996, which - fun fact - was the SoC that this driver
(or rather SDE, its downstream origin) was meant for and first tested on.

It has some hardware that differs from the modern SoCs, so not a lot of
current structs could have been reused. It's also seemingly the only SoC
supported by DPU that uses RGB pipes.

Signed-off-by: Konrad Dybcio <konrad.dybcio@somainline.org>